Backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
openai
psycopg2-binary
python-dotenv
pandas
scikit-learn